feat: Update YouTubePlayer and TimeCounter components
The YouTubePlayer component has been updated to include additional playerVars for autoplay, mute, controls, disablekb, enablejsapi, fs, iv_load_policy, modestbranding, and rel. This improves the functionality and user experience of the YouTube player.

The TimeCounter component has been modified to conditionally render the StopButton based on the value of the formdone variable. If formdone is true, the StopButton will be displayed with the text "중도포기하기". If formdone is false, the StopButton will be displayed with the text "바로 시작하기". This enhances the user interface of the TimeCounter component.

Please remove any unnecessary details and meta information from the commit message.